pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "file-triage"
version = "0.1.0"
description = "Toolkit for analysing and organising chaotic file and folder collections across multiple disks."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
# Optional: content-based MIME detection for text preview (e.g. no extension / wrong extension)
preview = [
    "python-magic>=0.4.27",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.scripts]
file-triage = "file_triage.cli:main"
